// projects/cntt/src/app/models/menuheader.ts
export interface MenuHeader {
  icon1: String;
  name1: String;
  url1: String;
  sub1: Array<Object>;
  icon2: String;
  name2: String;
  url2: String;
  sub2: Array<Object>;
  icon3: String;
  name3: String;
  url3: String;
  sub3: Array<Object>;
  icon4: String;
  name4: String;
  url4: String;
  sub4: Array<Object>;
  icon5: String;
  name5: String;
  url5: String;
  sub5: Array<Object>;
}